microservice/samrequirements.txt
fastapi
uvicorn
pillow
segment_anything
numpy
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
pydantic